move code to create private user registration to its own function
