+++ examples/interwallet/src/cryptography.ts
@@ -1,17 +1,18 @@
 import nacl from "tweetnacl"
 
 import { equals } from "uint8arrays"
-import { bytesToHex, hexToBytes, recoverTypedDataAddress } from "viem/utils"
+import { bytesToHex, getAddress, hexToBytes, keccak256, recoverTypedDataAddress } from "viem/utils"
 
 import * as Messages from "./protocols/messages"
 
 import { KeyBundle, PrivateUserRegistration, PublicUserRegistration } from "./interfaces"
+import { WalletClient } from "wagmi"
 
 export const getPublicUserRegistration = ({ privateKey: _, ...user }: PrivateUserRegistration) => user
 
 export const getRegistrationKey = (address: string) => `/interwallet/v0/registration/${address}`
 
-export const buildMagicString = (pin: string) => `[Password: ${pin}]
+const buildMagicString = (pin: string) => `[Password: ${pin}]
 
 Generate a new messaging key?
 
@@ -19,7 +20,7 @@ Signing this message will allow the application to read & write messages from yo
 
 Only do this when setting up your messaging client or mobile application.`
 
-export function constructTypedKeyBundle(keyBundle: KeyBundle) {
+function constructTypedKeyBundle(keyBundle: KeyBundle) {
 	const types = {
 		EIP712Domain: [{ name: "name", type: "string" }],
 		KeyBundle: [
@@ -90,5 +91,24 @@ export function assert(condition: unknown, message?: string): asserts condition 
 		return
 	} else {
 		throw new Error(message ?? "assertion error")
+	}
+}
+
+export const createPrivateUserRegistration = async (
+	walletClient: WalletClient,
+	userAddress: string,
+	pin: string
+): Promise<PrivateUserRegistration> => {
+	const magicString = buildMagicString(pin)
+	const signature = await walletClient.signMessage({ message: magicString })
+	const privateKey = keccak256(signature)
+	const keyBundle = makeKeyBundle(privateKey)
+	const typedKeyBundle = constructTypedKeyBundle(keyBundle)
+	const keyBundleSignature = await walletClient.signTypedData(typedKeyBundle)
+	return {
+		address: getAddress(userAddress),
+		privateKey,
+		keyBundle,
+		keyBundleSignature,
 	}
 }
